backend/requirements.txt
flask==2.0.1
flask-cors==3.0.10
pillow==9.0.0
numpy==1.22.0
ultralytics>=8.0.0
torch>=1.7.0
torchvision>=0.8.0
